Update view edit y delete
Update view edit y delete

+++ controller/PublicacionController.php
@@ -43,14 +43,24 @@ class PublicacionController
 
     public function delete()
     {
-        $id = !empty($_REQUEST["id"]) ? htmlentities($_REQUEST["id"]) : "";
-        $exito = $this->model->delete($id);
-        $this->redirectWithMessage(
-            $exito,
-            "Publicacion eliminada exitosamente",
-            "No se pudo realizar la eliminación",
-            "index.php?c=publicacion&f=index"
-        );
+        if(!isset($_SESSION)){session_start();}
+        if(isset($_SESSION['usuario'])){
+            $usuario = $_SESSION['usuario'];
+            $rol = $usuario['idRolFK']; 
+            if ($rol == 1){
+                $id = !empty($_REQUEST["id"]) ? htmlentities($_REQUEST["id"]) : "";
+                $exito = $this->model->delete($id);
+                $this->redirectWithMessage(
+                    $exito,
+                    "Publicacion eliminada exitosamente",
+                    "No se pudo realizar la eliminación",
+                    "index.php?c=publicacion&f=index");
+            }else if($rol != 1){
+                $_SESSION["mensaje"] = "Accion no permitida para tu perfil";
+                $_SESSION["color"] = "danger";
+                header("Location: index.php?c=publicacion&f=index");
+            }
+        }
     }
 
     public function redirectWithMessage($exito, $exitoMsg, $errMsg, $redirectUrl)
@@ -105,23 +115,30 @@ class PublicacionController
     }
 
     public function view_edit(){
-        if ($rol == 1){
-            $id = htmlentities($_GET["id"]);
-            $publi = $this->model->selectOne($id);
-            if($publi==null){
-                $_SESSION["mensaje"] = "No se pudo encontrar la publicación a editar";
+        if(!isset($_SESSION)){session_start();}
+        if(isset($_SESSION['usuario'])){
+            $usuario = $_SESSION['usuario'];
+            $rol = $usuario['idRolFK']; 
+            if ($rol == 1){
+                $id = htmlentities($_GET["id"]);
+                $publi = $this->model->selectOne($id);
+                
+                if($publi==null){
+                    $_SESSION["mensaje"] = "No se pudo encontrar la publicación a editar";
+                    $_SESSION["color"] = "danger";
+                    header("Location: index.php?c=publicacion&f=index");
+                }
+                $modeloTipoPublicacion = new TipoPublicacionDAO();
+                $tiposPublicaciones = $modeloTipoPublicacion->selectAll("");
+                $modeloPrioridad = new PrioridadDAO();
+                $prioridades = $modeloPrioridad->selectAll("");
+                $titulo = "Editar publicación";
+                require_once VPUBLICACIONES . 'edit.php';
+            }else if($rol != 1){
+                $_SESSION["mensaje"] = "Accion no permitida para tu perfil";
                 $_SESSION["color"] = "danger";
                 header("Location: index.php?c=publicacion&f=index");
             }
-            $modeloTipoPublicacion = new TipoPublicacionDAO();
-            $tiposPublicaciones = $modeloTipoPublicacion->selectAll("");
-            $modeloPrioridad = new PrioridadDAO();
-            $prioridades = $modeloPrioridad->selectAll("");
-            $titulo = "Editar publicación";
-            require_once VPUBLICACIONES . 'edit.php';
-        }else if($rol != 1){
-            echo "Acción no permitida para tu perfil";
-            header("Location: index.php?c=publicacion&f=index");
         }
     }
